+++ utilities/autoDelete.js
@@ -15,7 +15,7 @@ module.exports = new CronJob('00 00 00 * * *', () => {
             if (err) return err;
             link.remove({ _id: id })
           })
-          if (data[entry].imageEmbed) {
+          if (data[entry].imageEmbed && fs.existsSync(`./public/assets/img/${data[entry].URL}.jpeg`)) {
             fs.unlinkSync(`./public/assets/img/${data[entry].URL}.jpeg`)
           }
         } catch (err) {
